+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 * Fix sending emails
 * Greatly improve load time by displaying page sooner
 * Fix blank web overlays
+* Revert back to m.facebook.com
 
 ## v1.7.8
 * Emergency fix for notifications